refactor(tests): share duplicated helpers via tests/common
